compare against usize

+++ crates/sequencer/src/commitment/strategy.rs
@@ -121,7 +121,7 @@ where
         let compressed_state_diff = compress_blob(&borsh::to_vec(&merged_state_diff)?);
 
         // Threshold is checked by comparing compressed state diff size as the data will be compressed before it is written on DA
-        let state_diff_threshold_reached = compressed_state_diff.len() as u64 > MAX_TXBODY_SIZE;
+        let state_diff_threshold_reached = compressed_state_diff.len() > MAX_TXBODY_SIZE;
 
         if state_diff_threshold_reached {
             self.last_state_diff.clone_from(&l2_state_diff);
